requirements.txt
requests
psycopg2-binary
openai
python-dotenv
